+++ src/qnethernet/QNEthernetClient.cpp
@@ -68,8 +68,9 @@ int EthernetClient::connect(const IPAddress ip, const uint16_t port) {
 int EthernetClient::connect(const char* const host, const uint16_t port) {
 #if LWIP_DNS
   IPAddress ip;
-  if (!DNSClient::getHostByName(host, ip,
-                                QNETHERNET_DEFAULT_DNS_LOOKUP_TIMEOUT)) {
+  if (!DNSClient::getHostByName(
+          host, ip,
+          static_cast<uint32_t>(QNETHERNET_DEFAULT_DNS_LOOKUP_TIMEOUT))) {
     // INVALID_SERVER (-2)
     return false;
   }
@@ -98,8 +99,9 @@ bool EthernetClient::connectNoWait(const char* const host,
                                    const uint16_t port) {
 #if LWIP_DNS
   IPAddress ip;
-  if (!DNSClient::getHostByName(host, ip,
-                                QNETHERNET_DEFAULT_DNS_LOOKUP_TIMEOUT)) {
+  if (!DNSClient::getHostByName(
+          host, ip,
+          static_cast<uint32_t>(QNETHERNET_DEFAULT_DNS_LOOKUP_TIMEOUT))) {
     // INVALID_SERVER (-2)
     return false;
   }
